consumer_reddit/requirements.txt
kafka-python
pandas
pyarrow
s3fs